Fix inventory stock parsing from Int to Float
- Changed parseInt to parseFloat for currentStock in create/update APIs
- Added step='0.01' to stock quantity input in InventoryModal
- Now supports fractional quantities (e.g., 42ml, 2.5kg)
- Fixes inventory items not showing/loading issue

+++ app/api/inventory/[id]/route.ts
@@ -120,7 +120,7 @@ export async function PUT(
 
     // Handle stock change
     if (currentStock !== undefined) {
-      const newStock = parseInt(currentStock)
+      const newStock = parseFloat(currentStock)
       if (newStock !== currentItem.currentStock) {
         updates.currentStock = newStock
         logs.push({
